Split widget: overview panel
